completed chapter 17; working arithmatic calculator

+++ clox/src/chunk.cpp
@@ -18,10 +18,10 @@ void Chunk::writeByte(const uint8_t byte, const size_t line)
 }
 
 
-uint8_t Chunk::addConstant(const Value value)
+int Chunk::addConstant(const Value value)
 {
 	constants.write(value);
-	return static_cast<uint8_t>(constants.size() - 1);
+	return static_cast<int>(constants.size() - 1);
 }
 
 void Chunk::disassemble(const char* name) const
